requirements.txt
Django==4.1.7
django-crispy-forms==2.1
spacy==3.7.2
en_core_web_sm==3.7.2
scikit-learn==1.2.1
pandas==2.1.1
joblib==1.2.0
djangorestframework==3.14.0